backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for foodgram service"
authors = ["Кирюшин Виталий <kiryuschin.vitalij@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.6"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
djoser = "^2.2.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
flake8 = "^6.1.0"
psycopg2-binary = "^2.9.9"
python-levenshtein = "^0.22.0"
django-colorfield = "^0.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
